@@ -15,12 +15,15 @@
 //     along with this program.  If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
+using Microsoft.Extensions.Options;
+
 namespace slskd.Messaging.API
 {
     using System.Threading.Tasks;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Mvc;
+    using slskd.Relay;
     using Soulseek;
 
     /// <summary>
@@ -33,21 +36,28 @@ namespace slskd.Messaging.API
     [Consumes("application/json")]
     public class PublicChatController : ControllerBase
     {
-        public PublicChatController(ISoulseekClient soulseekClient)
+        public PublicChatController(ISoulseekClient soulseekClient, IOptionsSnapshot<Options> optionsSnapshot)
         {
             Client = soulseekClient;
+            OptionsSnapshot = optionsSnapshot;
         }
 
         private ISoulseekClient Client { get; }
+        private IOptionsSnapshot<Options> OptionsSnapshot { get; }
 
         /// <summary>
         ///     Starts public chat.
         /// </summary>
         /// <returns></returns>
         [HttpPost]
-        [Authorize]
+        [Authorize(Policy = AuthPolicy.Any)]
         public async Task<IActionResult> Start()
         {
+            if (Program.IsRelayAgent)
+            {
+                return Forbid();
+            }
+
             await Client.StartPublicChatAsync();
             return StatusCode(StatusCodes.Status201Created);
         }
@@ -57,9 +67,14 @@ namespace slskd.Messaging.API
         /// </summary>
         /// <returns></returns>
         [HttpDelete]
-        [Authorize]
+        [Authorize(Policy = AuthPolicy.Any)]
         public async Task<IActionResult> Stop()
         {
+            if (Program.IsRelayAgent)
+            {
+                return Forbid();
+            }
+
             await Client.StopPublicChatAsync();
             return StatusCode(StatusCodes.Status204NoContent);
         }
