@@ -35,8 +35,10 @@ namespace slskd.Shares
         /// <param name="localPath"></param>
         /// <param name="raw"></param>
         /// <param name="remotePath"></param>
+        /// <param name="directories"></param>
+        /// <param name="files"></param>
         [JsonConstructor]
-        public Share (string id, string alias, bool isExcluded, string localPath, string raw, string remotePath)
+        public Share (string id, string alias, bool isExcluded, string localPath, string raw, string remotePath, int? directories, int? files)
         {
             Id = id;
             Alias = alias;
@@ -44,6 +46,8 @@ namespace slskd.Shares
             LocalPath = localPath;
             Raw = raw;
             RemotePath = remotePath;
+            Directories = directories;
+            Files = files;
         }
 
         /// <summary>
@@ -87,5 +91,13 @@ namespace slskd.Shares
         public string LocalPath { get; init; }
         public string Raw { get; init; }
         public string RemotePath { get; init; }
+        public int? Directories { get; private set; }
+        public int? Files { get; private set; }
+
+        public void UpdateStatistics(int directories, int files)
+        {
+            Directories = directories;
+            Files = files;
+        }
     }
 }
