@@ -69,7 +69,7 @@ namespace slskd.Transfers
         /// <param name="username">The username associated with the file.</param>
         /// <param name="filename">The filename of the file for which the position is to be estimated.</param>
         /// <returns>The estimated queue position of the file.</returns>
-        /// <exception cref="FileNotFoundException">Thrown if the specified filename is not enqueued.</exception>
+        /// <exception cref="NotFoundException">Thrown if the specified filename is not enqueued.</exception>
         int EstimatePosition(string username, string filename);
 
         /// <summary>
@@ -105,11 +105,11 @@ namespace slskd.Transfers
             Configure(OptionsMonitor.CurrentValue);
         }
 
+        private int GlobalSlots { get; set; } = 0;
         private Dictionary<string, UploadGroup> Groups { get; set; } = new Dictionary<string, UploadGroup>();
         private int LastGlobalSlots { get; set; }
         private string LastOptionsHash { get; set; }
         private ILogger Log { get; } = Serilog.Log.ForContext<UploadQueue>();
-        private int MaxSlots { get; set; } = 0;
         private IOptionsMonitor<Options> OptionsMonitor { get; }
         private SemaphoreSlim SyncRoot { get; } = new SemaphoreSlim(1, 1);
         private ConcurrentDictionary<string, List<Upload>> Uploads { get; set; } = new ConcurrentDictionary<string, List<Upload>>();
@@ -251,7 +251,7 @@ namespace slskd.Transfers
         /// <param name="username">The username associated with the file.</param>
         /// <param name="filename">The filename of the file for which the position is to be estimated.</param>
         /// <returns>The estimated queue position of the file.</returns>
-        /// <exception cref="FileNotFoundException">Thrown if the specified filename is not enqueued.</exception>
+        /// <exception cref="NotFoundException">Thrown if the specified filename is not enqueued.</exception>
         public int EstimatePosition(string username, string filename)
         {
             var groupName = Users.GetGroup(username);
@@ -269,7 +269,7 @@ namespace slskd.Transfers
                 // find this user's uploads
                 if (!Uploads.TryGetValue(username, out var uploadsForUser))
                 {
-                    throw new FileNotFoundException($"File {filename} is not enqueued for user {username}");
+                    throw new NotFoundException($"File {filename} is not enqueued for user {username}");
                 }
 
                 // find the position of the requested file in the user's queue
@@ -280,7 +280,7 @@ namespace slskd.Transfers
 
                 if (localPosition < 0)
                 {
-                    throw new FileNotFoundException($"File {filename} is not enqueued for user {username}");
+                    throw new NotFoundException($"File {filename} is not enqueued for user {username}");
                 }
 
                 // start the position to the local position within this user's queue; the user's own files must be completed
@@ -321,7 +321,7 @@ namespace slskd.Transfers
 
             if (globalPosition < 0)
             {
-                throw new FileNotFoundException($"File {filename} is not enqueued for user {username}");
+                throw new NotFoundException($"File {filename} is not enqueued for user {username}");
             }
 
             return globalPosition;
@@ -378,7 +378,7 @@ namespace slskd.Transfers
                     return;
                 }
 
-                MaxSlots = options.Global.Upload.Slots;
+                GlobalSlots = options.Global.Upload.Slots;
 
                 // statically add built-in groups
                 var groups = new List<UploadGroup>()
@@ -392,7 +392,7 @@ namespace slskd.Transfers
                     {
                         Name = Application.PrivilegedGroup,
                         Priority = 0,
-                        Slots = MaxSlots,
+                        Slots = GlobalSlots,
                         UsedSlots = GetExistingUsedSlotsOrDefault(Application.PrivilegedGroup),
                         Strategy = QueueStrategy.FirstInFirstOut,
                     },
@@ -400,17 +400,17 @@ namespace slskd.Transfers
                     {
                         Name = Application.DefaultGroup,
                         Priority = options.Groups.Default.Upload.Priority,
-                        Slots = options.Groups.Default.Upload.Slots,
+                        Slots = Math.Min(options.Groups.Default.Upload.Slots, GlobalSlots),
                         UsedSlots = GetExistingUsedSlotsOrDefault(Application.DefaultGroup),
-                        Strategy = (QueueStrategy)Enum.Parse(typeof(QueueStrategy), options.Groups.Default.Upload.Strategy, true),
+                        Strategy = options.Groups.Default.Upload.Strategy.ToEnum<QueueStrategy>(),
                     },
                     new UploadGroup()
                     {
                         Name = Application.LeecherGroup,
                         Priority = options.Groups.Leechers.Upload.Priority,
-                        Slots = options.Groups.Leechers.Upload.Slots,
+                        Slots = Math.Min(options.Groups.Leechers.Upload.Slots, GlobalSlots),
                         UsedSlots = GetExistingUsedSlotsOrDefault(Application.LeecherGroup),
-                        Strategy = (QueueStrategy)Enum.Parse(typeof(QueueStrategy), options.Groups.Leechers.Upload.Strategy, true),
+                        Strategy = options.Groups.Leechers.Upload.Strategy.ToEnum<QueueStrategy>(),
                     },
                 };
 
@@ -419,9 +419,9 @@ namespace slskd.Transfers
                 {
                     Name = kvp.Key,
                     Priority = kvp.Value.Upload.Priority,
-                    Slots = kvp.Value.Upload.Slots,
+                    Slots = Math.Min(kvp.Value.Upload.Slots, GlobalSlots),
                     UsedSlots = GetExistingUsedSlotsOrDefault(kvp.Key),
-                    Strategy = (QueueStrategy)Enum.Parse(typeof(QueueStrategy), kvp.Value.Upload.Strategy, true),
+                    Strategy = kvp.Value.Upload.Strategy.ToEnum<QueueStrategy>(),
                 }));
 
                 Groups = groups.ToDictionary(g => g.Name);
@@ -442,7 +442,11 @@ namespace slskd.Transfers
 
             try
             {
-                if (Groups.Values.Sum(g => g.UsedSlots) >= MaxSlots)
+                // if the total number of used slots across all groups is greater than or equal to the configured
+                // number of global slots, just exit. we *could* proceed, but uploads would stack up behind the
+                // global semaphore in Soulseek.NET and we wouldn't be able to control the order in which those
+                // were processed, so don't do that.
+                if (Groups.Values.Sum(g => g.UsedSlots) >= GlobalSlots)
                 {
                     return null;
                 }
