@@ -4,8 +4,9 @@ import CodeMirror from '@uiw/react-codemirror';
 import { StreamLanguage, syntaxHighlighting, defaultHighlightStyle  } from '@codemirror/language';
 import { yaml } from '@codemirror/legacy-modes/mode/yaml';
 
-const CodeEditor = ({ value, onChange = () => {}, ...rest}) => {
+const CodeEditor = ({ value, theme, onChange = () => {}, ...rest}) => {
   console.log(defaultHighlightStyle);
+
   return (
     <CodeMirror
       value={value}
@@ -13,6 +14,7 @@ const CodeEditor = ({ value, onChange = () => {}, ...rest}) => {
         StreamLanguage.define(yaml),  
         syntaxHighlighting(defaultHighlightStyle, {fallback: true})]}
       onChange={(value, _) => onChange(value)}
+      theme={theme}
       { ...rest}
     />
   );
