@@ -10,8 +10,9 @@ import Logs from './Logs';
 import Options from './Options';
 import Shares from './Shares';
 import { Switch } from '../Shared';
+import Data from './Data';
 
-const System = ({ state = {}, options = {} }) => {
+const System = ({ state = {}, theme, options = {} }) => {
   const { params: { tab }, ...route } = useRouteMatch();
   const history = useHistory();
 
@@ -27,7 +28,7 @@ const System = ({ state = {}, options = {} }) => {
         </Switch>
         Info
       </Menu.Item>), 
-      render: () => <Tab.Pane><Info state={state}/></Tab.Pane>, 
+      render: () => <Tab.Pane><Info state={state} theme={theme}/></Tab.Pane>, 
     },
     {
       route: 'options', 
@@ -36,7 +37,7 @@ const System = ({ state = {}, options = {} }) => {
         icon: 'options', 
         content: 'Options', 
       }, 
-      render: () => <Tab.Pane className='full-height'><Options options={options} /></Tab.Pane>,
+      render: () => <Tab.Pane className='full-height'><Options options={options} theme={theme}/></Tab.Pane>,
     },
     {
       route: 'shares', 
@@ -48,7 +49,16 @@ const System = ({ state = {}, options = {} }) => {
         </Switch>
         Shares
       </Menu.Item>),
-      render: () => <Tab.Pane><Shares state={state.shares}/></Tab.Pane>,
+      render: () => <Tab.Pane><Shares state={state.shares} theme={theme}/></Tab.Pane>,
+    },
+    {
+      route: 'data',
+      menuItem: {
+        key: 'data',
+        icon: 'database',
+        content: 'Data',
+      },
+      render: () => <Tab.Pane className='full-height'><Data theme={theme}/></Tab.Pane>,
     },
     { 
       route: 'logs', 
