@@ -1,5 +1,6 @@
 const urlBase = (window.urlBase === '/' ? '' : window.urlBase) || '';
-const rootUrl = process.env.NODE_ENV === 'production' ? urlBase : `http://localhost:5000${urlBase}`;
+const devPort = window.port ?? 5030;
+const rootUrl = process.env.NODE_ENV === 'production' ? urlBase : `http://localhost:${devPort}${urlBase}`;
 const apiBaseUrl = `${rootUrl}/api/v0`;
 const hubBaseUrl = `${rootUrl}/hub`;
 const tokenKey = 'slskd-token';
