@@ -49,6 +49,13 @@
         {
             var (queue, _) = GetFixture(new Options()
             {
+                Global = new Options.GlobalOptions 
+                {
+                    Upload = new Options.GlobalOptions.GlobalUploadOptions
+                    {
+                        Slots = int.MaxValue,
+                    },
+                },
                 Groups = new Options.GroupsOptions()
                 {
                     Default = new Options.GroupsOptions.BuiltInOptions()
@@ -79,6 +86,13 @@
         {
             var (queue, _) = GetFixture(new Options()
             {
+                Global = new Options.GlobalOptions
+                {
+                    Upload = new Options.GlobalOptions.GlobalUploadOptions
+                    {
+                        Slots = int.MaxValue,
+                    },
+                },
                 Groups = new Options.GroupsOptions()
                 {
                     Leechers = new Options.GroupsOptions.LeecherOptions()
@@ -110,6 +124,13 @@
         {
             var (queue, _) = GetFixture(new Options()
             {
+                Global = new Options.GlobalOptions
+                {
+                    Upload = new Options.GlobalOptions.GlobalUploadOptions
+                    {
+                        Slots = int.MaxValue,
+                    },
+                },
                 Groups = new Options.GroupsOptions()
                 {
                     UserDefined = new Dictionary<string, Options.GroupsOptions.UserDefinedOptions>()
@@ -168,6 +189,13 @@
             {
                 var options = new Options()
                 {
+                    Global = new Options.GlobalOptions
+                    {
+                        Upload = new Options.GlobalOptions.GlobalUploadOptions
+                        {
+                            Slots = int.MaxValue,
+                        },
+                    },
                     Groups = new Options.GroupsOptions()
                     {
                         UserDefined = new Dictionary<string, Options.GroupsOptions.UserDefinedOptions>()
@@ -209,6 +237,63 @@
                 Assert.Equal(group, p.Name);
                 Assert.Equal(priority, p.Priority);
                 Assert.Equal(slots, p.Slots);
+                Assert.Equal(0, p.UsedSlots);
+                Assert.Equal(strategy, p.Strategy);
+            }
+
+            [Theory, AutoData]
+            public void Limits_Group_Slots_To_Global_Slot_Count(string group, int priority, QueueStrategy strategy)
+            {
+                var options = new Options()
+                {
+                    Global = new Options.GlobalOptions
+                    {
+                        Upload = new Options.GlobalOptions.GlobalUploadOptions
+                        {
+                            Slots = 42,
+                        },
+                    },
+                    Groups = new Options.GroupsOptions()
+                    {
+                        UserDefined = new Dictionary<string, Options.GroupsOptions.UserDefinedOptions>()
+                        {
+                            {
+                                group,
+                                new Options.GroupsOptions.UserDefinedOptions()
+                                {
+                                    Upload = new Options.GroupsOptions.UploadOptions()
+                                    {
+                                        Priority = priority,
+                                        Slots = int.MaxValue, // lots
+                                        Strategy = strategy.ToString(),
+                                    }
+                                }
+                            },
+                        }
+                    }
+                };
+
+                // do not pass options; init with defaults
+                var (queue, mocks) = GetFixture();
+
+                var groups = queue.GetProperty<Dictionary<string, UploadGroup>>("Groups");
+
+                // user defined group does not exist
+                Assert.False(groups.ContainsKey(group));
+
+                // reconfigure
+                mocks.OptionsMonitor.RaiseOnChange(options);
+
+                // get the new copy
+                groups = queue.GetProperty<Dictionary<string, UploadGroup>>("Groups");
+
+                Assert.True(groups.ContainsKey(group));
+
+                var p = groups[group];
+
+                Assert.Equal(group, p.Name);
+                Assert.Equal(priority, p.Priority);
+                Assert.Equal(42, p.Slots); // clamped to global value
                 Assert.Equal(0, p.UsedSlots);
                 Assert.Equal(strategy, p.Strategy);
             }
@@ -540,7 +625,7 @@
 
                 queue.SetProperty("Uploads", uploads);
 
-                var result = queue.InvokeMethod<Upload>("Process");
+                _ = queue.InvokeMethod<Upload>("Process");
 
                 var groups = queue.GetProperty<Dictionary<string, UploadGroup>>("Groups");
 
